refactor(macro): move logic target
Also remove DLL magic.
