+++ lib/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(about)
+add_subdirectory(macro)
