+++ lib/macro/macro.cpp
