+++ lib/macro/parse.hpp
@@ -17,10 +17,12 @@
 
 #pragma once
 
-#include <QtCore/QtGlobal>
+#include <QtCore>
+#include <tuple>
 
-#if defined(MACRO_LIBRARY)
-  #define MACRO_EXPORT Q_DECL_EXPORT
-#else
-  #define MACRO_EXPORT Q_DECL_IMPORT
-#endif
+namespace macro {
+// Analyze a macro's text body, and attempt to extract header information.
+// Tuple returns 1) is the header well formed, 2) what is the macro's name,
+// 3) how many arguments does the macro require?
+std::tuple<bool, QString, quint8> analyze_macro_definition(QString macro_text);
+}; // End namespace macro
