+++ lib/macro/registered.cpp
