+++ lib/macro/registry.cpp
