+++ test/macro/CMakeLists.txt
@@ -1,0 +1,2 @@
+file(GLOB_RECURSE tests CONFIGURE_DEPENDS "${CMAKE_CURRENT_LIST_DIR}/*.cpp")
+add_pepp_test(TARGET macro TESTS ${tests})
