+++ test/macro/registry.cpp
